+++ pkg/query/model/model.go
@@ -200,10 +200,16 @@ func (sr *StreamResult) CopyFrom(tmp, other *StreamResult) bool {
 		heap.Push(h, other)
 	}
 
+	seenElementIDs := make(map[uint64]bool)
+
 	// Pop from heap to build tmp with topN
 	for h.Len() > 0 && tmp.Len() < tmp.topN {
 		res := heap.Pop(h).(*StreamResult)
-		tmp.CopySingleFrom(res)
+		elementID := res.ElementIDs[res.idx]
+		if !seenElementIDs[elementID] {
+			seenElementIDs[elementID] = true
+			tmp.CopySingleFrom(res)
+		}
 		res.idx++
 		if res.idx < res.Len() {
 			heap.Push(h, res)
@@ -298,10 +304,15 @@ func MergeStreamResults(results []*StreamResult, topN int, asc bool) *StreamResu
 	}
 
 	mergedResult := NewStreamResult(topN, asc)
+	seenElementIDs := make(map[uint64]bool)
 
 	for h.Len() > 0 && mergedResult.Len() < topN {
 		sr := heap.Pop(h).(*StreamResult)
-		mergedResult.CopySingleFrom(sr)
+		elementID := sr.ElementIDs[sr.idx]
+		if !seenElementIDs[elementID] {
+			seenElementIDs[elementID] = true
+			mergedResult.CopySingleFrom(sr)
+		}
 		sr.idx++
 		if sr.idx < sr.Len() {
 			heap.Push(h, sr)